Fixed bug due to reusing PageLayout instance on Route change
When Reconciling after a route change, PageLayout form a different page would re-render and React would 'think' it was the same instance, so the hooks were not called the same way as the page controllers would be different. By having a different key prop for each PageLayout, we ensure that every time the application re-render due to a page change, a different PageLayout will be rendered.